virtiofsd: add initial support for shared versions
Not backward compatible with previous kernels, so please only use with
kernel that has version table support (this will need to be cleaned up).

No READDIRPLUS support in the kernel for versioned entries, so disable for
now.

Attribute timeout is set to "infinity", so changes to underlying filesystem
won't be visible.  This also needs to be fixed, but is best for testing the
versioning since the shared version is the only thing that will force
refreshing metadata and dcache lookups.

No caching metadata modifications yet.

Start "ireg" daemon before starting any fuse servers.

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Fix by:
Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>