virtiofsd: Add the low level fuse functions for fsnotify
Add the low level fuse functions in virtiofsd corresponding to fsnotify
request and fsnotify notification.

The low level request servicing function parses the data within the
"fuse_notify_fsnotify_in" struct that is part of the fsnotify request sent
by the a guest. The data contains the aggregated mask for the events that
are placed on the inode in the guest. Then it calls the fsnotify
corresponding handler within passthrough_ll.c

The low level fsnotify notification function does the exact opposite, it
sends the event data about an inode to the guest through the notification
queue. Specifically, it sends the mask related to an event, the inode
number, the generation of the inode, the parent of the monitored inode,
a file name associated with the inode (if available) and
a boolean to instruct the guest to notify both the monitored inode and its
parent.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>