virtiofsd: Add the fsnotify related structs in virtiofsd
Add a new struct "fuse_fsnotify" in "fuse_i.h" that holds the information
about the fsnotify instances (created by virtiofsd). Add the struct to the
"fuse_session" struct since "fuse_fsnotify" needs to be globaly accessible
by the fsnotify specific thread.

The use of a dedicated fsnotify thread is mandatory so that virtiofsd has
the resources to service guest requests but at the same time to monitor for
fsnotify events.

Also create a hashtable "fh_to_inode" which holds the mappings between file
handles to inodes. The purpose of this hashtable is to easily map file
handles to inodes when an fsnotify event occurs and virtiofsd needs to send
a notification to the guest. We specifically use file handles for the
mappings since the event notification backend is implemented with fanotify
in the unprivileged file handle (fid) mode.

Currently the server will only create one global fsnotify/fanotify instance
to monitor for events. However, for future implementations we introduce a
list in "fuse_fsnotify" which can hold more than one fsnotify/fanotify
instances and even inotify instances.

Finally, virtiofsd polls the fsnotify instance (descriptor) for events
using the epoll subsystem.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>