virtiofsd: Add the fsnotify thread that monitors for events
Add the dedicated thread that monitors the fsnotify events. The thread is
constantly polling (with epoll) the fsnotify instance (descriptor)
created by virtiofsd.

The thread is created during the initialization function of virtiofsd. If
the guest kernel wants remote fsnotify notification enabled and virtiofsd
supports it then virtiofsd creates the fsnotify dedicated thread.

When an event is pending for the fsnotify instance, the inotify thread
invokes the event handler which reads the event, maps the event to the
corresponding inode and then sends the information to the guest.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>