virtiofsd: Cleanup fsnotify thread when virtiofsd exits or reinitializes
During the main virtqueue thread cleanup, wait for the fsnotify thread to
exit safely during the virtiofsd exit or its reinitialization.

The fsnotify thread will check if the member-variable "cleanup_fsnotify" of
the struct "fuse_fsnotify" under struct "fuse_session". The "fuse_session"
struct is globaly accessible and it is tied to the life-time of virtiofsd.
Since the fsnotify thread is not tied to the life time of a virtqueue but
on the life time of virtiofsd, the use of the of "cleanup_fsnotify" member
of struct "fuse_fsnotify" seems a logical way to notify the fsnotify thread
about any impending exit or re-initialization of virtiofsd.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>