virtiofsd: Add the functions that service the fsnotify requests
Add the passthrough functions that service the fsnotify requests. A
fsnotify request can implement two actions: 1) Remove a watch, 2)
Add/Modify and existing watch.

In any case, virtiofsd will first check if there is an existing fsnotify
instance corresponding to the guest process that sent the request. In the
case that the request refers to the deletion of a watch without an existing
fsnotify instance, an error is generated. Else virtiofsd will create a new
fsnotify instance and will then create a watch on the inode targeted by the
request.

Since fsnotify needs the name of the file to add a watch to it, something
that is not available directly in virtiofsd, readlinkat is used along with
the open proc descriptor of the target inode to find a "link" name to said
inode. Then virtiofsd uses the "fanotify_mark" system call to add a
watch on the target inode (file or dir).

There is also a reference counter attached to the fsnotify instance
created by virtiofsd. With each addition of a watch related to the
fsnotify instance the reference counter is increased by one and when a
watch is removed it is decreased by one. When the counter reaches zero
then the fsnotify instance is deleted. This helps to prevent the fsnotify
instance being alive even after the guest process associated with it exits.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>