virtiofsd: Set the size for the notification event message
Add the struct "fuse_notify_fsnotify_out" to the union "fuse_notify_union"
in fuse_virtio.c. Then round up the whole notification size to 512.

This size should be plently for now for sending the notification queue
events, however this is temporary solution. The problem is that along with
a notification event we might have to send the filename related to that
event. The filename can have a variable size though.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>