virtiofsd: Store the inode file handle in the lo_inode struct
Add two additional members to the lo_inode struct in passthrough_ll.c.
These two members are the file handle for the inode and the current
aggregated mask (for events) for that inode.

We need the file handle for the inode when we get the events from the host
to match the file handle stored in the event (for the fanotify fid mode) to
the inode it refers to.

The aggregated mask is needed to keep a state of what watch operations
(add/remove) have happened for the specific inode.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>