virtiofsd: Add the mapping functions for inodes and their file handles
In lo_do_lookup when a new inode is created then get its file handle and
store it in the mapping hashtables. When we get a notification event from
the host (fanotify fid mode) then we are able to match the file handle
returned in the event with the ones that we have in our mapping hashtables
and then get the related inode.

Also in the lo_forget_one when an inode is deleted by viritofsd remove the
mapping from its file handle to the inode from the hashtable.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>